requirements.txt
async-lru==1.0.2
pyaes==1.6.1
Pyrogram==1.2.9
PySocks==1.7.1
pytgvoip==0.0.7.1
pytgvoip-pyrogram==0.0.11
